+++ src/lib/switchbot/client.ts
