+++ src/lib/switchbot/zodSchema.ts
